PC-605: move eligibility tests out to their own files
helps break up the output from pytest
