+++ tests/eligibility/test_scenario_25.py
@@ -1,0 +1,18 @@
+import pytest
+
+from help_to_heat.frontdoor.eligibility import country_council_tax_bands
+from tests.test_eligibility import countries_england_scotland_wales, own_property_tenant_landlord, \
+    property_types_not_park_home, yes_no, epc_DEU, no, household_income_less_than_31k, _run_scenario, accept_both
+
+
+@pytest.mark.parametrize("country", countries_england_scotland_wales)
+@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
+@pytest.mark.parametrize("property_type", property_types_not_park_home)
+@pytest.mark.parametrize("park_home_main_residence", yes_no)
+@pytest.mark.parametrize("epc_rating", epc_DEU)
+def test_scenario_25(country, own_property, property_type, park_home_main_residence, epc_rating):
+    benefits = no
+    household_income = household_income_less_than_31k
+    for council_tax_band in country_council_tax_bands[country]["eligible"]:
+        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
+                             benefits, household_income) == accept_both
