+++ tests/test_eligibility.py
@@ -2,10 +2,8 @@ import datetime
 import unittest
 import uuid
 
-import pytest
-
 from help_to_heat.frontdoor import interface
-from help_to_heat.frontdoor.eligibility import country_council_tax_bands, calculate_eligibility
+from help_to_heat.frontdoor.eligibility import calculate_eligibility
 from help_to_heat.frontdoor.mock_os_api import MockOSApi
 from help_to_heat.portal import models
 
@@ -39,436 +37,9 @@ property_type_park_home = "Park home"
 property_types_not_park_home = ("House", "Bungalow", "Apartment, flat or maisonette")
 property_types = property_types_not_park_home + (property_type_park_home,)
 
-accept_both = ("ECO4", "GBIS")
+accept_both = ("GBIS", "ECO4")
 accept_gbis = ("GBIS",)
 deny = ()
-
-
-@pytest.mark.parametrize("country", countries_not_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epcs)
-# a country not in england or wales should fail no matter what else is input
-# however testing all other possibilities is too much, so varying only a subset
-def test_scenario_1(country, property_type, council_tax_band, epc_rating):
-    assert _run_scenario(country, own_property_owner, property_type, yes, council_tax_band, epc_rating,
-                         yes, household_income_less_than_31k) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epcs)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_2(country, council_tax_band, epc_rating, benefits, household_income):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = no
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_3(country, council_tax_band, epc_rating, benefits, household_income):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_4(country, council_tax_band, epc_rating, household_income):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = yes
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_5(country, council_tax_band, epc_rating):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = no
-    household_income = household_income_less_than_31k
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_6(country, council_tax_band, epc_rating):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = no
-    household_income = household_income_more_than_31k
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_7(country, property_type, park_home_main_residence, epc_rating, benefits, household_income):
-    own_property = own_property_owner
-
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_8(country, property_type, park_home_main_residence, epc_rating, household_income):
-    own_property = own_property_owner
-    benefits = yes
-
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_9(country, property_type, park_home_main_residence, epc_rating):
-    own_property = own_property_owner
-    benefits = no
-    household_income = household_income_less_than_31k
-
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_10(country, property_type, park_home_main_residence, epc_rating):
-    own_property = own_property_owner
-    benefits = no
-    household_income = household_income_more_than_31k
-
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_11(country, property_type, park_home_main_residence, epc_rating, benefits, household_income):
-    own_property = own_property_owner
-
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_14(country, property_type, park_home_main_residence, epc_rating, household_income):
-    own_property = own_property_owner
-    benefits = yes
-
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_12(country, property_type, park_home_main_residence, epc_rating):
-    own_property = own_property_owner
-    benefits = no
-    household_income = household_income_less_than_31k
-
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_13(country, property_type, park_home_main_residence, epc_rating):
-    own_property = own_property_owner
-    benefits = no
-    household_income = household_income_more_than_31k
-
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epcs)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_15(country, own_property, council_tax_band, epc_rating, benefits, household_income):
-    property_type = property_type_park_home
-    park_home_main_residence = no
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_16(country, own_property, council_tax_band, epc_rating, benefits, household_income):
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_17(country, own_property, council_tax_band, epc_rating, household_income):
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = yes
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_19(country, own_property, council_tax_band, epc_rating):
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = no
-    household_income = household_income_less_than_31k
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_18(country, own_property, council_tax_band, epc_rating):
-    own_property = own_property_owner
-    property_type = property_type_park_home
-    park_home_main_residence = yes
-    benefits = no
-    household_income = household_income_more_than_31k
-
-    assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_20(country, own_property, property_type, park_home_main_residence, epc_rating, benefits, household_income):
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_FG)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_22(country, own_property, property_type, park_home_main_residence, epc_rating, household_income):
-    benefits = yes
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_FG)
-def test_scenario_23(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = yes
-    household_income = household_income_less_than_31k
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_FG)
-def test_scenario_21(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = yes
-    household_income = household_income_more_than_31k
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_24(country, own_property, property_type, park_home_main_residence, epc_rating, household_income):
-    benefits = yes
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEU)
-def test_scenario_25(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = yes
-    household_income = household_income_less_than_31k
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEU)
-def test_scenario_26(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = yes
-    household_income = household_income_more_than_31k
-    for council_tax_band in country_council_tax_bands[country]["eligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_gbis
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_27(country, own_property, property_type, park_home_main_residence, epc_rating, benefits, household_income):
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_28(country, own_property, property_type, park_home_main_residence, epc_rating, household_income):
-    benefits = yes
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_29(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = no
-    household_income = household_income_less_than_31k
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == accept_both
-
-
-@pytest.mark.parametrize("country", countries_england_scotland_wales)
-@pytest.mark.parametrize("own_property", own_property_tenant_landlord)
-@pytest.mark.parametrize("property_type", property_types_not_park_home)
-@pytest.mark.parametrize("park_home_main_residence", yes_no)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-def test_scenario_30(country, own_property, property_type, park_home_main_residence, epc_rating):
-    benefits = no
-    household_income = household_income_more_than_31k
-    for council_tax_band in country_council_tax_bands[country]["ineligible"]:
-        assert _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating,
-                             benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_not_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_ABC)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_31(country, property_type, council_tax_band, epc_rating, benefits, household_income):
-    assert _run_scenario(country, own_property_owner, property_type, yes, council_tax_band, epc_rating,
-                         benefits, household_income) == deny
-
-
-@pytest.mark.parametrize("country", countries_not_england_scotland_wales)
-@pytest.mark.parametrize("council_tax_band", council_tax_bands)
-@pytest.mark.parametrize("epc_rating", epc_DEFGU)
-@pytest.mark.parametrize("benefits", yes_no)
-@pytest.mark.parametrize("household_income", household_incomes)
-def test_scenario_32(country, property_type, council_tax_band, epc_rating, benefits, household_income):
-    assert _run_scenario(country, own_property_owner, property_type, yes, council_tax_band, epc_rating,
-                         benefits, household_income) == accept_both
 
 
 def _run_scenario(country, own_property, property_type, park_home_main_residence, council_tax_band, epc_rating, benefits, household_income):
